% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_assistant_files.R
\name{assistant_delete_file}
\alias{assistant_delete_file}
\title{Delete File from Assistant}
\usage{
assistant_delete_file(assistant_name, file_id)
}
\arguments{
\item{assistant_name}{Name of the assistant}

\item{file_id}{UUID of the file to delete}
}
\value{
List with http response, content, and status_code.
  - 200: The file has been successfully deleted
  - 404: File not found
  - 500: Internal error
}
\description{
Deletes a file from an assistant.
}
\examples{
\dontrun{
assistant_delete_file("my-assistant", "070513b3-022f-4966-b583-a9b12e0290ff")
}
}
