% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_index_operations.R
\name{configure_index}
\alias{configure_index}
\title{Configure index in terms of replicas and pod size}
\usage{
configure_index(index_name, replicas, pod_type, controller = "controller")
}
\arguments{
\item{replicas}{integer
The desired number of replicas for the index.}

\item{pod_type}{The new pod type for the index. One of s1, p1, or p2 appended with . and one of x1, x2, x4, or x8.}

\item{controller}{}
}
\description{
Configure index in terms of replicas and pod size
}
