% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_collection_operations.R
\name{create_collection}
\alias{create_collection}
\title{Create Collection}
\usage{
create_collection(name, source)
}
\arguments{
\item{name}{Name for the new collection}

\item{source}{Name of the source index to create the collection from}
}
\value{
List with http response, content, and status_code
- 201: The collection has been successfully created
- 400: Bad request (quota exceeded or invalid name)
- 409: A collection with this name already exists
- 500: Internal error
}
\description{
Creates a collection from an existing index. Collections are static copies
of an index that can be used to create new indexes.
}
\examples{
\dontrun{
create_collection("my-collection", source = "my-index")
}
}
