% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_index_operations.R
\name{create_index}
\alias{create_index}
\title{Create Index}
\usage{
create_index(
  name,
  dimension,
  metric,
  pods,
  replicas,
  pod_type,
  metadata_config,
  source_collection
)
}
\arguments{
\item{source_collection}{}
}
\description{
Create Index
}
