% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_collection_operations.R
\name{delete_collection}
\alias{delete_collection}
\title{Delete Collection}
\usage{
delete_collection(collection_name)
}
\arguments{
\item{collection_name}{Name of the collection to delete}
}
\value{
List with http response, content, and status_code
- 202: The collection has been successfully deleted
- 404: Collection not found
- 500: Internal error
}
\description{
Deletes an existing collection.
}
\examples{
\dontrun{
delete_collection("my-collection")
}
}
