% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_collection_operations.R
\name{describe_collection}
\alias{describe_collection}
\title{Describe collection}
\usage{
describe_collection(collection_name, controller = "controller")
}
\arguments{
\item{collection_name}{}
}
\description{
Describe collection
}
