% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_index_operations.R
\name{describe_index}
\alias{describe_index}
\title{Describe Index}
\usage{
describe_index(index_name, controller = "controller")
}
\arguments{
\item{return_controller}{}
}
\value{
200 Configuration information and deployment status of the index
404 Index not found
500 Internal error. Can be caused by invalid parameters.
}
\description{
Describe Index
}
