% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_vector_operations.R
\name{describe_index_stats}
\alias{describe_index_stats}
\title{Descrive Index stats}
\usage{
describe_index_stats(controller)
}
\arguments{
\item{controller}{}
}
\description{
Descrive Index stats
}
