% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_connect.R
\name{extract_vector_controller}
\alias{extract_vector_controller}
\title{Get controller (DEPRECATED)}
\usage{
extract_vector_controller(index, vector = "svc")
}
\arguments{
\item{index}{Index description object}

\item{vector}{Ignored}
}
\value{
Host string
}
\description{
This function is deprecated. Use `get_index_host()` instead.
}
