% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_connect.R
\name{get_index_host}
\alias{get_index_host}
\title{Get Index Host}
\usage{
get_index_host(index)
}
\arguments{
\item{index}{Index description object from describe_index()}
}
\value{
Full host string for data plane API calls
}
\description{
Extracts the full host from an index description for data plane operations.
}
\examples{
\dontrun{
index <- describe_index("my-index")
host <- get_index_host(index)
}
}
