% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_vector_operations.R
\name{vector_delete}
\alias{vector_delete}
\title{Title}
\usage{
vector_delete(
  index,
  ids = NULL,
  delete_all = FALSE,
  name_space = "",
  filter = NULL
)
}
\arguments{
\item{filter}{}
}
\description{
Title
}
