% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_vector_operations.R
\name{vector_query}
\alias{vector_query}
\title{Query Vectors}
\usage{
vector_query(
  index,
  vector,
  top_k = 5,
  filter = list(),
  name_space = "",
  include_metadata = TRUE,
  include_values = FALSE,
  tidy = TRUE
)
}
\arguments{
\item{tidy}{}
}
\description{
Query Vectors
}
