% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_vector_operations.R
\name{vector_upsert}
\alias{vector_upsert}
\title{Upsert}
\usage{
vector_upsert(index, embeddings, vector_id, meta_data = "", name_space = "")
}
\arguments{
\item{name_space}{}
}
\description{
Upsert
}
