% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_collection_operations.R
\name{create_collection}
\alias{create_collection}
\title{Create Collection}
\usage{
create_collection(name, source_collection)
}
\arguments{
\item{source_collection}{}
}
\value{
list with api call and content.
201 The collection has been successfully created.
400 Bad request. Request exceeds quota or collection name is invalid.
409 A collection with the name provided already exists.
500 Internal error. Can be caused by invalid parameters.
}
\description{
Create Collection
}
