% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_collection_operations.R
\name{delete_collection}
\alias{delete_collection}
\title{Delete collection}
\usage{
delete_collection(collection_name, controller = "controller")
}
\arguments{
\item{collection_name}{}
}
\value{
202 The index has been successfully deleted.
404 Collection not found.
500 Internal error. Can be caused by invalid parameters.
}
\description{
Delete collection
}
