% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_index_operations.R
\name{delete_index}
\alias{delete_index}
\title{Delete Index}
\usage{
delete_index(index_name)
}
\arguments{
\item{index_name}{Name of the index to delete}
}
\value{
List with http response, content, and status_code
- 202: The index has been successfully deleted
- 404: Index not found
- 500: Internal error
}
\description{
Deletes an existing index.
}
\examples{
\dontrun{
delete_index("my-index")
}
}
