% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_vector_operations.R
\name{describe_index_stats}
\alias{describe_index_stats}
\title{Descrive Index stats}
\usage{
describe_index_stats(index_name = NA, project_name = NA)
}
\arguments{
\item{project_name}{}
}
\description{
Descrive Index stats
}
