% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_connect.R
\name{extract_vector_controller}
\alias{extract_vector_controller}
\title{Get controller}
\usage{
extract_vector_controller(index, vector = "svc")
}
\arguments{
\item{vector}{}
}
\description{
Get controller
}
