% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_vector_operations.R
\name{vector_update}
\alias{vector_update}
\title{Update Vectors}
\usage{
vector_update(
  index,
  embeddings = NULL,
  vector_id,
  meta_data = "",
  name_space = ""
)
}
\arguments{
\item{name_space}{}
}
\description{
Update Vectors
}
